Issue #328 - Add docker compose example file
This patch adds a docker-compose example file for helping new
users to use bookkeeper.

It runs by default an ensemble with 3 bookies and a client
application based in the following tutorial, but you can leverage docker
compose features to scale up the number of instances.

https://bookkeeper.apache.org/archives/docs/master/bookkeeperTutorial.html

Ref: Issue #328

Signed-off-by: Antonio Ojea <antonio.ojea.garciagmail.com>

Author: Antonio Ojea <antonio.ojea.garcia@gmail.com>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #871 from aojea/docker, closes #328