ISSUE #931,#907: Add option to track task execution time
Fixes a bug in OrderedScheduler introduced in
e33ec10aa400f32c2e0278c15ea80a0f624e5919 which failed to track execution
time with some calls and adds an option to enable it in the bookie.  Also
fixes a bug with task_queued duration.

Add a simple mock for remembering stats long enough to verify that
counters are actually used and sensible in unit tests and bake it into
BookKeeperClusterTestCase so that we can write tests to ensure that the
stats are actually counted and make sense.  Use said mock to add simple
tests for top level read and write stats validating this fix.

(bug W-4276826)
(bug W-4268290)
Signed-off-by: Samuel Just <sjustsalesforce.com>

Author: Samuel Just <sjust@salesforce.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #932 from athanatos/forupstream/issue-931, closes #931, closes #907