Build Docker image from maven
The docker image is build on the install phase, using the tarball
generated by the bookkeeper-dist-server module.

It is only built if the docker profile is enabled.
```
mvn install -Pdocker
```

This removes the need to update the docker file after each release, as
the image is automatically generated as part of the release process.