Add descriptive message for TooManyRequestsException
After upgrading from 4.5 to 4.6 I found a lot of "Unexpected condition" in some benchmarks.

TooManyRequestsException is a new error introduced in 4.6, there is not description, so the default description is "Unexpected condition", without a debugger it is not possible to understand the cause of the error.

Author: Enrico Olivelli <eolivelli@apache.org>
Author: eolivelli <eolivelli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #921 from eolivelli/toomanyrequests-message