ISSUE #986: in LocalBK.main do Sys.exit in the case of an Exc
Descriptions of the changes in this PR:

in LocalBookKeeper.main method if there is any exception then it should do System.exit. This is needed,  because, some non-daemon thread (probably in ZK or some other dependent service) is preventing the JVM from exiting, though there is exception in main thread.

Master Issue: #986

Author: cguttapalem <cguttapalem@salesforce.com>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #987 from reddycharan/minorlocalbookkeeperfix, closes #986