Increase page size for fetching github labels
Descriptions of the changes in this PR:

*Problem*

The default page size for github api is 30 items. The number of labels at bookkeeper are more than 30 items, so `type/task` is not showing up at merge script.

*Solution*

This change is to bump the page size to 100 items.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>

This closes #1013 from sijie/increase_page_size