Test upgrade from old server version to new version
This patch contains two tests.
- Upgrade between all released version, up to latest version
- Upgrade direct from 4.1.0 to latest version

In all cases the old client should continue to work with the new version.
In one case(4.2->4.3), we broken client compatibility so the the new client
stops working with the old server.

These tests replace and expand on the testing done by
TestBackwardCompat#testCompat410 with the old BC tests.

Master Issue: #903

Author: Ivan Kelly <ivank@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1028 from ivankelly/compat401