BP-25: MovingChecksumToProto--Refactor the checksum part of bookkeeper
Descriptions of the changes in this PR:

Current the checksum implementation is in client module while the checksum semantic is more close to protocol. Moreover, moving the checksum implementation to protocol will avoid server module's dependency to client module when doing checksum in server side.

Master Issue: #1007

Author: Arvin <arvindevel@gmail.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>, Venkateswararao Jujjuri <vjujjuri@salesforce.com>

This closes #1008 from ArvinDevel/BP_25_MovingChecksumToProto