URLDecode key/cert paths in TestTLS
In TestTLS we load keys/certs from the classpath, which means we get
the filename as a URL. Since it is a URL it is URL encoded for special
characters. Usually this is not an issue, but in the case that there
are special characters in the path, these will be encoded, and then
when we try to open the file we get a FileNotFoundException. This
occurs on jenkins, when there are two parallel builds on the same
job (it adds a N to the workspace name), and was causing all TLS
tests to fail in that case.

Author: Ivan Kelly <ivank@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1074 from ivankelly/tls-fix