Consolidate the protobuf definition used in state store and in client/service (#38)
state store and client/service were started as two different projects, and were doing in parallel. so there are some duplication and overlaps on defining the k/v operations.
now these two projects were put together, it would be good to consolidate the protobuf definition to have it only exist in one place.