Split all versions images into released and latest
The released tarballs rarely change, while the latest changes every
build. This patch splits the all version image into a base images with
the released tarballs, and a image which includes all released and the
very latest tarball.

This allows the released image to be cached/pushed to dockerhub, which
will in turn allow us to speed up our CI builds that use the all
version image.