Issue#1152 stats for durability violations in write/read path
Durability contract demands that there are WQ copies spread
across fault zones as per the ensemble placement policy.
There are lots of areas where this contract is violated.

Add stats to track two areas that are well:
a. Write Path: if the write fails after satisfying AQ
b. Read Path: If the checksum error is detected on the read path

Signed-off-by: Venkateswararao Jujjuri (JV) <vjujjuri@salesforce.com>