Issue #1152: stats for durability violations in write/read path
Descriptions of the changes in this PR:

Durability contract demands that there are WQ copies spread
across fault zones as per the ensemble placement policy.
There are lots of areas where this contract is violated.

Add stats to track two areas that are well:
a. Write Path: if the write fails after satisfying AQ
b. Read Path: If the checksum error is detected on the read path

Signed-off-by: Venkateswararao Jujjuri (JV) <vjujjurisalesforce.com>

Master Issue: #1152

Author: JV Jujjuri <vjujjuri@salesforce.com>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1155 from jvrao/bk-ssue-1152-stats-1, closes #1152