Replace DoubleByteBuf with ByteBufList
This PR aims at a better approach to fix the issue reported in #1108

Instead of using `CompositeByteBuf` which has the downside of needing multiple object allocations per each usage (~7-8), we should have a way to aggregate multiple buffers into a single "logical" frame with no additional overhead.

This PR introduces a `ByteBufList` class which is just a holder of a list of buffers, without imitating the `ByteBuf` API. You can pass a `ByteBufList` to the channel and have the encoder take care of the multiple writes.

Instances of ByteBufList are ref-counted and released to a pool after the write is completed. There are no object allocations or memory copies involved.

Author: Matteo Merli <mmerli@apache.org>
Author: Sijie Guo <guosijie@gmail.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Venkateswararao Jujjuri (JV) <None>

This closes #1141 from merlimat/bytebuf-list-master