Use BookieServer#getLocalAddress as the identifier in the test
*Problem*

Issue #1097 introduced using `loopback` nic as the bookie server identifier. However a few test cases construct
the bookie socket address using `InetAddress.getLocalHost()` and the bookie port. This bookie socket address can
be different from the actual socket address that a bookie is advertising to zookeeper. It will cause test cases
fail with some network settings.

*Solution*

This change use `BookieServer#getLocalAddress()` to retrieve the actual bookie socket address. so it would respect
to the server configuration and use the right socket address in the test.