Introduce "IGNORE IT CI" to ignore integration tests ci result
Descriptions of the changes in this PR:

*Problem*

Integration tests ci is not very reliable at this moment. so most of the time we have to "IGNORE CI" to allow changes
to be merged. However "IGNORE CI" is a bit too ambiguous, which can potentially ignore other CI results.

*Solution*

This change introduces "IGNORE IT CI" to ignore integration tests ci results only.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Ivan Kelly <ivank@apache.org>, Enrico Olivelli <eolivelli@gmail.com>

This closes #1173 from sijie/ignore_integration_tests