V2 decoder marks the reads index of the payload on write
This means that any storage implementation that receives that payload,
can reset the reader index, without jumping into bytes it knows
nothing about.

This was the source of an issue when using V2 protocol with
DbLedgerStorage where the data could not be successfully read back
from the ledger.

Author: Ivan Kelly <ivank@apache.org>

Reviewers: Matteo Merli <mmerli@apache.org>, Sijie Guo <sijie@apache.org>

This closes #1180 from ivankelly/broken-v2-rocks