Test that we don't break compat with Yahoo custom version
We've been merging a lot of changes from yahoo's internal bookkeeper
fork in the past months. The plan is that once everything is in, yahoo
will move to using stock opensource bookkeeper. We need to ensure that
this move goes smoothly, otherwise they wouldn't move. So this set of
tests checks whether master will work with what they are running
internally.