Issue #1164: TestTruncate#testPurgeLogs is flaky
Descriptions of the changes in this PR:

*Problem*

log segments are truncated in the background after TTL. the truncation is done asynchronously.
so when we finish writing all the log records, the truncation might still not complete yet.

*Solution*

Improve the test case to wait until the last truncation attempt is completed before asserting the log segments

Master Issue: #1164

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1165 from sijie/fix_test_truncate, closes #1164