Fix GSSAPIBookKeeperTest
Descriptions of the changes in this PR:

*Problem*

in GSSAPIBookKeeperTest, `InetAddress.getLocalHost().getHostName()` was used for adding principle to kdc keytab. However in the tests, it uses advertisedAddress (introduced by #1097).

So it causes authentication failed because server can't be found.

*Solution*

Change to use `Bookie. getBookieAddress(..)` to get the real hostname used by bookie for adding keytab entries.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Charan Reddy Guttapalem <reddycharan18@gmail.com>, Enrico Olivelli <eolivelli@gmail.com>

This closes #1191 from sijie/fix_gssapibookkeeper_test