Added MockBookKeeper for unit tests
This mock class is what we use in Pulsar to do unit tests for the managed-ledger library.

This approach might be useful in general, in all cases when creating unit tests of an application that uses the BookKeeper API.

The `MockBookKeeper` behaves like the regular BK client, except it needs no bookies or zookkeeper. All ledgers and data are just kept in an hashmap in memory.

It is possible to inject failure in the sync/async calls, even after N steps. This is very useful to trigger failure in complex operations and cover this paths in unit tests.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1209 from merlimat/mock-bookkeeper