use FutureUtils/FutureEventListener in bookkeeper-common for dlog
Descriptions of the changes in this PR:

This is a sub-task for apache/bookkeeper#1024

After dlog is merged to bookkeeper, FutureUtils/FutureEventListener is duplicated at bookkeeper-common and distributedlog-common. This change is to remove FutureUtils/FutureEventListener from distributedlog-common and let dlog use the ones in bookkeeper-common.

The changes includes:

- remove `FutureUtils/FutureEventListener` from distributedlog-common
- change dlog classes to use the ones in bookkeeper-common
- change the dlog `OrderScheduler` to extend bookkeeper `OrderedScheduler` (required by FutureUtils.within).

Master Issue: #1024

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1208 from sijie/use_bookkeeper_futureutils