Pull as many reviewers as possible when merging
Github uses pagination on some apis. This means that if there are a
lot of comments on a issue, it may not pick up approvals by
default. This patch increases the number of review comments to pull,
from a default of 30 to 100. A proper fix would be to enable
pagination using the Link header, but that's a much bigger change, for
little return.