Fix TestBKSyncLogReader
*Problem*

The test case `TestBKSyncLogReader` is waiting an entry beyond a control record, where the number of control records can
vary depend on the timing of flush()/commit(). This introduces the flakiness of this test.

*Solution*

This change changes to wait for readahead go beyond the last user record, which is more deterministic.