Fix TestBKSyncLogReader
Descriptions of the changes in this PR:

*Problem*

The test case `TestBKSyncLogReader` is waiting an entry beyond a control record, where the number of control records can
vary depend on the timing of flush()/commit(). This introduces the flakiness of this test.

*Solution*

This change changes to wait for readahead go beyond the last user record, which is more deterministic.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Yiming Zang <yzang2016@gmail.com>, Jia Zhai <None>,  Philip Su <psu@twitter.com>

This closes #1168 from sijie/fix_test_sync_log_reader