Make table service client includes table service related jars
Descriptions of the changes in this PR:

This is a change for #1205

*Motivation*

in order to implify the other project to use table service, it would be good to make sure the client include table service related jars.

*Solution*

Include table service related jars in the client-all jar.

Master Issue: #1205

Author: Sijie Guo <sijie@apache.org>

Reviewers: Jia Zhai <None>

This closes #1226 from sijie/publish_table_service_jars