ISSUE #1229: PendingAddOp can get recycled before it gets executed
Descriptions of the changes in this PR:
*Problem*

The PendingAddOp can be recycled when it is cancelled before it is executed. so it will hit NPE when it is actually executed. This is a bug introduced by #1091

*Solution*

Only recycle PendingAddOp after it has been run.

Master Issue: #1229

Author: Sijie Guo <sijie@apache.org>

Reviewers: Andrey Yegorov <None>, Matteo Merli <mmerli@apache.org>, Venkateswararao Jujjuri (JV) <None>

This closes #1230 from sijie/fix_npe, closes #1229