Issue #570: getting rid of unnecessary synchronization in InterleavedLedgerStorage
Descriptions of the changes in this PR:

This is < sub-task1 > of Issue #570.

In InterleavedLedgerStorage, since the initial version of it, addEntry and processEntry methods are synchronized, but they are not required to be synchronized.

Removal of unnecessary synchronization in InterleavedLedgerStorage methods, as described in http://mail-archives.apache.org/mod_mbox/bookkeeper-dev/201707.mbox/%3CCAO2yDyZ946fp2S_qR2iL178hPiXgrnFGb%3DpvkyK4ReSYAtNLBw%40mail.gmail.com%3E

Master Issue: #570

Author: cguttapalem <cguttapalem@salesforce.com>

Reviewers: Ivan Kelly <ivank@apache.org>, Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #1225 from reddycharan/nonsynchaddinterleavedstorage, closes #570