Fixed Journal stats names
In `BOOKKEEPER-1009: Use multiple journals in bookie` 123eccd435a4a96a9147ed4a24efbe9025fe79ba there was a change in the metrics name that would be affecting also user not running with multiple journal.

It is a bit inconvenient to aggregate the stats in the metrics collector (how to aggregate 99pct latency for example).

I think the best option is to have a single metric even when using multiple journal threads.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Ivan Kelly <ivank@apache.org>, Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #1250 from merlimat/journal-metric-names