Jenkins seed job for testing DSLs
The job allows you to specify a repo and a branch. It will look for
all files named .test-infra/jenkins/jenkins_testing_job_*.groovy and
try to generate jobs for them. The seed job is in the folder
bookkeeper-jenkins-testing/ and has lookupStrategy set to
'SEED_JOB'. This means that any job generated by this seed will be
created in the same folder, thus lowering the chance of breaking our
existing production jobs.

Author: Ivan Kelly <ivank@apache.org>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #1248 from ivankelly/test-seed