Allow to send callbacks from journal thread
In BK 4.3 it was possible to configure 0 threads for the dispatching of journal callback threads.

There might be good reasons to not use a thread pool when dispatching the callback threads:
 * Avoiding one more context switch
 * Avoiding contention on the executor enqueuing
 * When using multiple journals, the journal thread is not a bottleneck anymore

We should allow the same convention (threads=0 means direct execution) that we have in other places.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Ivan Kelly <ivank@apache.org>, Sijie Guo <sijie@apache.org>

This closes #1258 from merlimat/journal-callbacks