Merge tests/jmh into microbenchmark
Descriptions of the changes in this PR:

tests/jmh was added a while ago, but it is got dropped out of root pom with mistakes.
mircobenchmark was added later. it is making sense to merge tests/jmh into microbenchmark module since they are using jmh to do microbenchmarks

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>

This closes #1265 from sijie/move_jmh_to_benchmarks