Fixed buffer release in v2 BookieProtocol
There was a buffers leak when using V2 bookie protocol. This was introduced with the refactoring and the merges from yahoo branch.

Note: this was only happening when selecting  v2 protocol in client configuration.

The current code was missing a release in both the add and read entries path. I've been running it for some time and the ref-counting seems correct after the fix.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1266 from merlimat/fix-refcounting