Set `ENABLE_DIGEST_TYPE_AUTODETECTION` to true as default value
Descriptions of the changes in this PR:

*Motivation*

We have recorded digest type in the ledger metadata since BK 4.5. it would be great to let client automatically detect the digest type
on opening ledger. So turn on the autodetection feature by default.

*Result*

digest type autodetection is on by default.

Author: Sijie Guo <sijie@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Jia Zhai <None>, Matteo Merli <mmerli@apache.org>

This closes #1252 from sijie/flip_default_values