Fixed DbStorage write cache segment index computation
When the write cache was merged from `yahoo-4.3` branch into master, an improvement went in around computing the the write cache segment index (doing shift instead of division) and that introduced a problem.

The write cache is broken into multiple segments (since buffers in JVM can only be at most 2GB). By default we use 1GB segments (or smaller if the write cache is overall smaller).

Because the shift count was wrong, the entries in the first segment were being ovrridden by entries in the 2nd segment (entries in the other segments were all fine).

That would lead to write some garbage in the entry log and subsequent read failures.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Sijie Guo <sijie@apache.org>

This closes #1279 from merlimat/fix-write-cache-segment-index