Refactored logAndConvertStatus() to avoid allocation when debug is off
The method `logAndConvertStatus(StatusCode status, int defaultStatus, Object... extraInfo)` is being called when processing the response for each write/read request in `PerChannelBookieClient`.

The logging is only done at debug level and the `extraInfo` is not used otherwise. By taking the `Object...` we're effectively taking a `Object[]` that is allocated each time, along with the boxing for ledgerId/entryId.

Refactored the code separate the logging and the status code conversion, so that we can skip the allocation when debug is off.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #1291 from merlimat/log-and-convert-status