Issue #1282: call to appendLedgersMap in flushCompactionLog
Descriptions of the changes in this PR:

with this change https://github.com/apache/bookkeeper/commit/3392beee5c70abe36d36604723e14d97b7764be9 compactionlog was introduced. But for compactionlog call to appendLedgersMap (// Append ledgers map at the end of entry log) isn't made.

So for these logs, getEntryLogMetadata call has to call extractEntryLogMetadataByScanning for getting the metadata of entrylog. Which is a perf hit.

Master Issue: #1282

Author: cguttapalem <cguttapalem@salesforce.com>

Reviewers: Yiming Zang <yzang2016@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #1283 from reddycharan/compactionlogfix, closes #1282