Use reflection based CRC32 to pass direct memory pointer
The crc update with the bytebuffer is very expensive, especially in netty > 4.1.12.

```java
crc.get().update(data.nioBuffer());
```

Converting a direct`ByteBuf` into a `ByteBuffer` with `data.nioBuffer()` is actually resorting to allocating an unpooled `DirectByteBuffer` which is actually killing the GC (since direct memory is freed when the `DirectByteBuffer` instances are GCed, but the pauses are > 1 second).

This is an adaptation of a change from Yahoo branch at https://github.com/yahoo/bookkeeper/commit/6c01ca2921a998f3bf2e85cacb27867773e7ea28

Initially I though this change was not needed anymore in 4.7 codebase, and that is true for the CRC32c variant. For the regular CRC32, though, we need to avoid the bad behavior of unpooled direct buffers.

Additionally, as in #1306, the thread local variable was made static to avoid the leakage of instances per each ledger.

Author: Matteo Merli <mmerli@apache.org>

Reviewers: Jia Zhai <None>, Sijie Guo <sijie@apache.org>

This closes #1307 from merlimat/crc32-bytebuffer and squashes the following commits:

8b81245a0 [Matteo Merli] Made classed package private
a7ff1573f [Matteo Merli] Use reflection based CRC32 to pass direct memory pointer